Conflicts:
	sql/src/main/java/io/crate/analyze/StatementAnalyzer.java
	sql/src/main/java/io/crate/planner/symbol/ByteLiteral.java
	sql/src/main/java/io/crate/planner/symbol/Literal.java
	sql/src/main/java/io/crate/planner/symbol/ShortLiteral.java
	sql/src/test/java/org/cratedb/integrationtests/SQLTypeMappingTest.java

+++ sql/src/main/java/io/crate/planner/symbol/ByteLiteral.java
@@ -28,7 +28,7 @@ import org.elasticsearch.common.io.stream.StreamOutput;
 
 import java.io.IOException;
 
-public class ByteLiteral extends NumberLiteral<Byte, ByteLiteral> {
+public class ByteLiteral extends NumberLiteral<Integer, ByteLiteral> {
 
     public static final SymbolFactory FACTORY = new SymbolFactory() {
         @Override
@@ -72,8 +72,8 @@ public class ByteLiteral extends NumberLiteral<Byte, ByteLiteral> {
     }
 
     @Override
-    public Byte value() {
-        return value;
+    public Integer value() {
+        return (int)value;
     }
 
     @Override
